% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{aggregate_to_annual}
\alias{aggregate_to_annual}
\title{Aggregate Daily to Annual}
\usage{
aggregate_to_annual(data)
}
\arguments{
\item{data}{Data frame with date and value columns}
}
\value{
Annual aggregated data
}
\description{
Aggregate Daily to Annual
}
