% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{aggregate_to_monthly}
\alias{aggregate_to_monthly}
\title{Aggregate Daily to Monthly}
\usage{
aggregate_to_monthly(data)
}
\arguments{
\item{data}{Data frame with date and value columns}
}
\value{
Monthly aggregated data
}
\description{
Aggregate Daily to Monthly
}
