% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosero_metrics_vis.R
\name{build_metrics_map_app}
\alias{build_metrics_map_app}
\title{Build the Shiny app mapping the problem catchments}
\usage{
build_metrics_map_app(poly)
}
\description{
Build the Shiny app mapping the problem catchments
}
\keyword{internal}
