% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosero_map.R
\name{check_map_viewer_packages}
\alias{check_map_viewer_packages}
\title{Check that suggested packages for the map viewer are installed}
\usage{
check_map_viewer_packages(need_simplify = FALSE)
}
\description{
Check that suggested packages for the map viewer are installed
}
\keyword{internal}
