% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosero_metrics_vis.R
\name{check_metrics_vis_packages}
\alias{check_metrics_vis_packages}
\title{Check that suggested packages for the metrics map are installed}
\usage{
check_metrics_vis_packages()
}
\description{
Check that suggested packages for the metrics map are installed
}
\keyword{internal}
