% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\docType{data}
\name{COLORS_DISCHARGE}
\alias{COLORS_DISCHARGE}
\title{Discharge plot colors}
\format{
An object of class \code{list} of length 2.
}
\usage{
COLORS_DISCHARGE
}
\description{
Discharge plot colors
}
\keyword{datasets}
