% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{COLORS_PRECIPITATION}
\alias{COLORS_PRECIPITATION}
\title{Precipitation plot colors (includes ET for seasonality panel)}
\usage{
COLORS_PRECIPITATION
}
\description{
Precipitation plot colors (includes ET for seasonality panel)
}
