% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\docType{data}
\name{COLORS_RUNOFF}
\alias{COLORS_RUNOFF}
\title{Runoff components plot colors}
\format{
An object of class \code{list} of length 5.
}
\usage{
COLORS_RUNOFF
}
\description{
Runoff components plot colors
}
\keyword{datasets}
