% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\docType{data}
\name{COLORS_WATER_BALANCE}
\alias{COLORS_WATER_BALANCE}
\title{Water balance plot colors}
\format{
An object of class \code{list} of length 6.
}
\usage{
COLORS_WATER_BALANCE
}
\description{
Water balance plot colors
}
\keyword{datasets}
