% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{copy_template_files}
\alias{copy_template_files}
\title{Copy Template Files to Project}
\usage{
copy_template_files(input_dir, overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{input_dir}{Destination input directory}

\item{overwrite}{Overwrite existing files?}

\item{quiet}{Suppress messages?}
}
\description{
Internal function to copy example/template input files to a new project.
}
\keyword{internal}
