% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{create_sobol_bounds}
\alias{create_sobol_bounds}
\title{Create Parameter Bounds Matrix for Sobol Sampling}
\usage{
create_sobol_bounds(par_bounds)
}
\arguments{
\item{par_bounds}{Tibble from load_parameter_bounds()}
}
\value{
Named tibble with min/max rows for sobol_matrices()
}
\description{
Create Parameter Bounds Matrix for Sobol Sampling
}
