% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosero_run.R
\name{detect_cosero_errors}
\alias{detect_cosero_errors}
\title{Check COSERO output for error patterns}
\usage{
detect_cosero_errors(output)
}
\description{
Check COSERO output for error patterns
}
\keyword{internal}
