% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{detect_timestep}
\alias{detect_timestep}
\title{Detect time step of input data}
\usage{
detect_timestep(data)
}
\arguments{
\item{data}{Data frame with Date column}
}
\value{
Character: "hourly", "daily", or "monthly"
}
\description{
Detect time step of input data
}
