% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{format_cosero_date}
\alias{format_cosero_date}
\title{Format date components to COSERO date string}
\usage{
format_cosero_date(date, hour = 0, minute = 0)
}
\arguments{
\item{date}{Date object}

\item{hour}{Hour (0-23)}

\item{minute}{Minute (0-59)}
}
\value{
String in format "YYYY MM DD HH MM"
}
\description{
Format date components to COSERO date string
}
