% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosero_run.R
\name{format_period_string}
\alias{format_period_string}
\title{Format COSERO date vector as readable string}
\usage{
format_period_string(settings)
}
\description{
Format COSERO date vector as readable string
}
\keyword{internal}
