% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{format_time_duration}
\alias{format_time_duration}
\title{Format Time Duration}
\usage{
format_time_duration(seconds)
}
\arguments{
\item{seconds}{Time in seconds}
}
\value{
Formatted string
}
\description{
Format Time Duration
}
\keyword{internal}
