% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosero_run.R
\name{handle_execution_failure}
\alias{handle_execution_failure}
\title{Handle execution failure reporting}
\usage{
handle_execution_failure(exec_result, runtime_seconds, project_path)
}
\description{
Handle execution failure reporting
}
\keyword{internal}
