% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosero_run.R
\name{handle_statevar_dmp}
\alias{handle_statevar_dmp}
\title{Handle statevar.dmp file location (silent)}
\usage{
handle_statevar_dmp(project_path)
}
\description{
Handle statevar.dmp file location (silent)
}
\keyword{internal}
