% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{modify_parameter_file}
\alias{modify_parameter_file}
\title{Modify COSERO Parameter File}
\usage{
modify_parameter_file(par_file, params, par_bounds, original_values)
}
\arguments{
\item{par_file}{Path to para.txt}

\item{params}{Named vector/list of sampled parameter values}

\item{par_bounds}{Parameter bounds table with modification types}

\item{original_values}{Original parameter values from file}
}
\description{
Modify COSERO Parameter File
}
