% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{modify_parameter_table}
\alias{modify_parameter_table}
\title{Modify COSERO Parameter File (Tabular Format)}
\usage{
modify_parameter_table(
  par_file,
  params,
  par_bounds,
  original_values,
  zones = NULL,
  quiet = FALSE
)
}
\arguments{
\item{par_file}{Path to parameter file}

\item{params}{Named vector/list of sampled parameter values}

\item{par_bounds}{Parameter bounds table with modification types}

\item{original_values}{Original parameter values from file}

\item{zones}{Vector of zone IDs to modify (NULL = all zones)}

\item{quiet}{Suppress messages}
}
\description{
Modify COSERO Parameter File (Tabular Format)
}
