% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{parse_cosero_date}
\alias{parse_cosero_date}
\title{Parse COSERO date string to components}
\usage{
parse_cosero_date(date_string)
}
\arguments{
\item{date_string}{String in format "YYYY MM DD HH MM" or vector c(YYYY, MM, DD, HH, MM)}
}
\value{
List with date, hour, minute
}
\description{
Parse COSERO date string to components
}
