% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{plot_seasonality_discharge}
\alias{plot_seasonality_discharge}
\title{Plot seasonality - Discharge}
\usage{
plot_seasonality_discharge(monthly_data)
}
\arguments{
\item{monthly_data}{Monthly aggregated discharge data}
}
\value{
A plotly object
}
\description{
Plot seasonality - Discharge
}
