% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{plotly_empty}
\alias{plotly_empty}
\title{Create empty plotly with message}
\usage{
plotly_empty(text = "No data available")
}
\arguments{
\item{text}{Message to display}
}
\value{
Plotly object
}
\description{
Create empty plotly with message
}
