% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosero_metrics_vis.R
\name{prepare_metrics_polygons}
\alias{prepare_metrics_polygons}
\title{Read, simplify and join the catchment polygons}
\usage{
prepare_metrics_polygons(
  catchments_shp,
  catchment_id_field,
  m,
  simplify_tolerance
)
}
\description{
Read, simplify and join the catchment polygons
}
\keyword{internal}
