% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{prepare_seasonality_data}
\alias{prepare_seasonality_data}
\title{Prepare seasonality data (monthly aggregation)}
\usage{
prepare_seasonality_data(subbasin_data, spinup_timesteps = 0)
}
\arguments{
\item{subbasin_data}{Output from prepare_subbasin_data()}

\item{spinup_timesteps}{Number of initial timesteps to exclude}
}
\value{
List with monthly aggregated data for each panel
}
\description{
Prepare seasonality data (monthly aggregation)
}
\details{
Fluxes (P, ET, Q) are summed, states (BW0, BW3, SWW) are averaged
}
