% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosero_run.R
\name{print_run_header}
\alias{print_run_header}
\title{Print the run header block}
\usage{
print_run_header(
  project_path,
  defaults_settings,
  statevar_source,
  tmmon_option
)
}
\description{
Print the run header block
}
\keyword{internal}
