% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosero_run.R
\name{print_run_results}
\alias{print_run_results}
\title{Print results summary with NSE/KGE table}
\usage{
print_run_results(output_data, runtime_seconds)
}
\description{
Print results summary with NSE/KGE table
}
\keyword{internal}
