% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{read_parameter_file}
\alias{read_parameter_file}
\title{Read Parameter Values from File}
\usage{
read_parameter_file(par_file, param_names)
}
\arguments{
\item{par_file}{Path to para.txt}

\item{param_names}{Vector of parameter names to read}
}
\value{
Named list of parameter values
}
\description{
Read Parameter Values from File
}
