% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosero_metrics_vis.R
\name{read_statistics_metrics}
\alias{read_statistics_metrics}
\title{Read statistics.txt and normalise it to subbasin / NSE / KGE / beta}
\usage{
read_statistics_metrics(stats_file)
}
\description{
Read statistics.txt and normalise it to subbasin / NSE / KGE / beta
}
\keyword{internal}
