% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosero_map.R
\name{resolve_sf_layer}
\alias{resolve_sf_layer}
\title{Resolve a path or sf object to an sf object}
\usage{
resolve_sf_layer(x, arg_name)
}
\description{
Resolve a path or sf object to an sf object
}
\keyword{internal}
