% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosero_metrics_vis.R
\name{resolve_statistics_file}
\alias{resolve_statistics_file}
\title{Resolve a project path or file path to statistics.txt}
\usage{
resolve_statistics_file(cosero_path)
}
\description{
Resolve a project path or file path to statistics.txt
}
\keyword{internal}
