% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartacus_preprocessing.R
\name{spartacus_preprocessing}
\alias{spartacus_preprocessing}
\title{SPARTACUS Data Preprocessing for COSERO}
\description{
Functions to extract SPARTACUS gridded climate data (precipitation and temperature)
and write them in COSERO meteorological input format.
}
\keyword{internal}
