% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{write_placeholder_png}
\alias{write_placeholder_png}
\title{Write a minimal placeholder PNG with error message}
\usage{
write_placeholder_png(filename)
}
\arguments{
\item{filename}{Path to write}
}
\description{
Write a minimal placeholder PNG with error message
}
\keyword{internal}
