% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\docType{data}
\name{COLORS_PRECIPITATION}
\alias{COLORS_PRECIPITATION}
\title{Precipitation plot colors (includes ET for seasonality panel)}
\format{
An object of class \code{list} of length 3.
}
\usage{
COLORS_PRECIPITATION
}
\description{
Precipitation plot colors (includes ET for seasonality panel)
}
\keyword{datasets}
