% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosero_readers.R
\name{detect_outputtype}
\alias{detect_outputtype}
\title{Detect COSERO output type}
\usage{
detect_outputtype(output_dir)
}
\arguments{
\item{output_dir}{Path to COSERO output directory}
}
\value{
Integer: 1, 2, or 3 corresponding to OUTPUTTYPE
}
\description{
Detect COSERO output type
}
