% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{plot_dotty}
\alias{plot_dotty}
\title{Create Dotty Plots}
\usage{
plot_dotty(
  parameter_sets,
  Y,
  y_label = "Output",
  n_col = 3,
  reference_line = NULL
)
}
\arguments{
\item{parameter_sets}{Parameter sets tibble}

\item{Y}{Output vector}

\item{y_label}{Y-axis label}

\item{n_col}{Number of columns in facet}

\item{reference_line}{Reference line value (optional)}
}
\value{
ggplot object
}
\description{
Create Dotty Plots
}
