% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{plot_ensemble_uncertainty}
\alias{plot_ensemble_uncertainty}
\title{Plot Ensemble Uncertainty}
\usage{
plot_ensemble_uncertainty(ensemble_output, observed = NULL)
}
\arguments{
\item{ensemble_output}{List of time series from extract_ensemble_output()}

\item{observed}{Observed data frame with date and value}
}
\value{
ggplot object
}
\description{
Plot Ensemble Uncertainty
}
