% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{plot_seasonality_precipitation}
\alias{plot_seasonality_precipitation}
\title{Plot seasonality - Precipitation + ET}
\usage{
plot_seasonality_precipitation(monthly_data)
}
\description{
Plot seasonality - Precipitation + ET
}
