% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_helpers.R
\name{plot_seasonality_runoff}
\alias{plot_seasonality_runoff}
\title{Plot seasonality - Runoff Components}
\usage{
plot_seasonality_runoff(monthly_data)
}
\arguments{
\item{monthly_data}{Monthly aggregated runoff components data}
}
\value{
A plotly object
}
\description{
Plot seasonality - Runoff Components
}
