% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{plot_sobol}
\alias{plot_sobol}
\title{Plot Sobol Indices}
\usage{
plot_sobol(sobol_indices, title = "Sobol Sensitivity Indices")
}
\arguments{
\item{sobol_indices}{Result from calculate_sobol_indices()}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Plot Sobol Indices
}
